Added HomeFragmentPage.java and Edited MianActivity.java, SignUp.java, and AndroidManifest.xml
Added HomeFragmentPage.java - basically connecting the home_fragment.xml to this file, so that i could go to that page. If you want to change the name of it, that's totally fine.
Edited MianActivity.java, SignUp.java, and AndroidManifest.xml

Now Sign In and Sign Up pages should works exactly as it suppose to. It should switch between sign up and sign in pages. It should Log you in, and It should sign you up.